ResourceModal refactored
