+++ frontend/src/redux/actions/modal.js
@@ -1,8 +1,23 @@
+import { modalEnum } from "../../utils/enums";
+
 export const MODAL_OPEN = "MODAL_OPEN";
 export const MODAL_CLOSE = "MODAL_CLOSE";
 export const MODAL_TOGGLE = "MODAL_TOGGLE";
 
 export const openModal = () => ({ type: MODAL_OPEN });
-export const openModalWithPayload = payload => ({ type: MODAL_OPEN, payload });
+export const openResourceModal = () => ({
+  type: MODAL_OPEN,
+  modalType: modalEnum.RESOURCE
+});
+export const openResourceModalWithPayload = payload => ({
+  type: MODAL_OPEN,
+  modalType: modalEnum.RESOURCE,
+  payload
+});
+export const openUserModalWithPayload = payload => ({
+  type: MODAL_OPEN,
+  modalType: modalEnum.USER,
+  payload
+});
 export const closeModal = () => ({ type: MODAL_CLOSE });
 export const toggleModal = () => ({ type: MODAL_TOGGLE });
