+++ frontend/src/utils/enums.js
@@ -4,4 +4,10 @@ const roleEnum = {
   PENDING: "PENDING"
 };
 
+const modalEnum = {
+  RESOURCE: "RESOURCE",
+  USER: "USER"
+};
+
 module.exports.roleEnum = roleEnum;
+module.exports.modalEnum = modalEnum;
